+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-12-08
+
+### Added
+
+- Ability to save investigation
+- Ability to load investigation
+- Ability to delete investigation
+- Ability to clear investigation data 
+
+### Fixed
+
+- GUI freezing due to absence of threaded transforms
+
 ## [1.0.0] - 2019-06-23
 
 ### Added